Add Ubuntu Configuration
This commit adds a playbook for configuring a new Ubuntu instance. The
only major modification is a new task for setting a static IP address,
which is different from Debian and Raspberry Pis.

Also, while testing this new playbook, I discovered that setting the
static ip too early in the process can cause problems if the playbook
takes too long to run. The new IP kicked in and the rest of the playbook
fails because it is searching for the wrong IP. To resolve this, I moved
the task for setting the IP to the bottom of the playbook for all the
configuration playbooks.