Update static IP tasks
There were a few hacky things in how I was setting static IPs for PIs
and Ubuntu. For Ubuntu it was just hardcoded spaces, which could
be solved by backrefs, and updating to the new netplan format for
default gateways. For Raspberry Pis I was getting the active network
interface name from the user, which is not ideal. I found a hard-to-read
but sufficient solution by getting the interface with the current IP
from the ansible facts.

Also there was a small bug where `locales` is not installed on Ubuntu,
so I couldn't generate the locales until that package was installed.